Refactor strategy names
