+++ test/strategy/anchor/fork/AnchorNonUSTStrategy.fork.test.ts
@@ -4,24 +4,27 @@ import { expect } from "chai";
 import { BigNumber, utils, constants } from "ethers";
 import {
   Vault,
-  USTStrategy,
+  AnchorNonUSTStrategy,
   MockERC20,
   MockChainlinkPriceFeed,
   MockERC20__factory,
   MockChainlinkPriceFeed__factory,
-} from "../../../typechain";
-import { generateNewAddress, ForkHelpers } from "../../shared/";
+} from "../../../../typechain";
+import { generateNewAddress, ForkHelpers } from "../../../shared/";
 import config from "./config.json";
 
-describe("USTStrategy Mainnet fork", () => {
+describe("AnchorNonUSTStrategy Mainnet fork", () => {
   let owner: SignerWithAddress;
   let alice: SignerWithAddress;
   let bob: SignerWithAddress;
 
   let vault: Vault;
-  let strategy: USTStrategy;
+  let strategy: AnchorNonUSTStrategy;
   let ustToken: MockERC20;
   let aUstToken: MockERC20;
+  let usdtToken: MockERC20;
+  let usdcToken: MockERC20;
+  let daiToken: MockERC20;
   // MockChainlinkPriceFeed has same interface as Mainnet, so we can use it for test
   let mockAUstUstFeed: MockChainlinkPriceFeed;
   const MIN_LOCK_PERIOD = 0; // set zero for test
@@ -36,22 +39,25 @@ describe("USTStrategy Mainnet fork", () => {
     await ForkHelpers.forkToMainnet(FORK_BLOCK);
     [owner, alice, bob] = await ethers.getSigners();
 
+    daiToken = new MockERC20__factory(owner).attach(config.dai);
+    usdcToken = new MockERC20__factory(owner).attach(config.usdc);
+    usdtToken = new MockERC20__factory(owner).attach(config.usdt);
     ustToken = new MockERC20__factory(owner).attach(config.ust);
     aUstToken = new MockERC20__factory(owner).attach(config.aUst);
 
     await ForkHelpers.mintToken(
-      ustToken,
+      usdtToken,
       alice,
-      utils.parseEther("100000000000")
+      utils.parseUnits("100000000000", 6)
     );
     await ForkHelpers.mintToken(
-      ustToken,
+      usdtToken,
       bob,
-      utils.parseEther("100000000000")
+      utils.parseUnits("100000000000", 6)
     );
   });
 
-  describe("Use Mainnet aUST / UST Chainlink feed", () => {
+  describe("Use Mainnet aUST / UST Chainlink feed (Underlying: USDT)", () => {
     before(async () => {
       mockAUstUstFeed = new MockChainlinkPriceFeed__factory(owner).attach(
         config.aUstUstFeed
@@ -59,15 +65,17 @@ describe("USTStrategy Mainnet fork", () => {
 
       const VaultFactory = await ethers.getContractFactory("Vault");
       vault = await VaultFactory.deploy(
-        ustToken.address,
+        usdtToken.address,
         MIN_LOCK_PERIOD,
         INVEST_PCT,
         owner.address
       );
 
-      const USTStrategyFactory = await ethers.getContractFactory("USTStrategy");
+      const AnchorNonUSTStrategyFactory = await ethers.getContractFactory(
+        "AnchorNonUSTStrategy"
+      );
 
-      strategy = await USTStrategyFactory.deploy(
+      strategy = await AnchorNonUSTStrategyFactory.deploy(
         vault.address,
         TREASURY,
         config.ethAnchorRouter,
@@ -75,23 +83,28 @@ describe("USTStrategy Mainnet fork", () => {
         ustToken.address,
         aUstToken.address,
         FEE_PCT,
-        owner.address
+        owner.address,
+        config.curve,
+        config.usdtI,
+        config.ustI
       );
 
       await strategy.connect(owner).grantRole(MANAGER_ROLE, owner.address);
 
+      await strategy.initializeStrategy(config.ustFeed, config.usdtFeed);
+
       await vault.setStrategy(strategy.address);
 
-      await ustToken
+      await usdtToken
         .connect(alice)
         .approve(vault.address, constants.MaxUint256);
-      await ustToken.connect(bob).approve(vault.address, constants.MaxUint256);
+      await usdtToken.connect(bob).approve(vault.address, constants.MaxUint256);
     });
 
     it("Mainnet fork (1)", async () => {
-      let amount = utils.parseEther("10000");
+      let amount = utils.parseUnits("10000", 6);
 
-      console.log(`Deposit ${utils.formatEther(amount)} UST by alice`);
+      console.log(`Deposit ${utils.formatUnits(amount, 6)} USDT by alice`);
       await vault.connect(alice).deposit({
         amount,
         claims: [
@@ -103,20 +116,26 @@ describe("USTStrategy Mainnet fork", () => {
         ],
         lockedUntil: 0,
       });
-      expect(await ustToken.balanceOf(vault.address)).to.be.equal(amount);
+      expect(await usdtToken.balanceOf(vault.address)).to.be.equal(amount);
       let exchangeRate = (await mockAUstUstFeed.latestRoundData()).answer;
       console.log("ExchangeRate: ", utils.formatEther(exchangeRate));
 
-      console.log(
-        `Invest: totalUnderlying - ${utils.formatEther(
-          await vault.totalUnderlying()
-        )} UST`
-      );
       await vault.connect(owner).updateInvested();
+      console.log(
+        `Invest: totalUnderlying - ${utils.formatUnits(
+          await vault.totalUnderlying(),
+          6
+        )} USDT`
+      );
+
       expect(await ustToken.balanceOf(vault.address)).to.be.equal("0");
       expect(await ustToken.balanceOf(strategy.address)).to.be.equal("0");
-      expect(await strategy.pendingDeposits()).to.be.equal(amount);
-      expect(await vault.totalUnderlying()).to.be.equal(amount);
+      expect(await usdtToken.balanceOf(vault.address)).to.be.equal("0");
+      expect(await usdtToken.balanceOf(strategy.address)).to.be.equal("0");
+      expect(await strategy.pendingDeposits()).to.be.equal(
+        "9984797021624030505513" // raw data in fork block
+      );
+      let totalUnderlying = BigNumber.from(await vault.totalUnderlying());
 
       let depositOperations = await strategy.depositOperations(0);
       let expectAUstReceive = utils.parseEther("8500");
@@ -132,18 +151,21 @@ describe("USTStrategy Mainnet fork", () => {
         expectAUstReceive
       );
 
-      let totalUnderlying = await vault.totalUnderlying();
       exchangeRate = (await mockAUstUstFeed.latestRoundData()).answer;
       console.log("ExchangeRate: ", utils.formatEther(exchangeRate));
 
+      totalUnderlying = await vault.totalUnderlying();
       console.log(
-        `FinishDepositStable: totalUnderlying - ${utils.formatEther(
-          totalUnderlying
-        )} UST (0 UST + ${utils.formatEther(expectAUstReceive)} aUST)`
+        `FinishDepositStable: totalUnderlying - ${utils.formatUnits(
+          totalUnderlying,
+          6
+        )} USDT (0 USDT + 0 UST + ${utils.formatEther(expectAUstReceive)} aUST)`
       );
 
-      amount = utils.parseEther("1000");
-      console.log(`Deposit ${utils.formatEther(amount)} UST by bob`);
+      expect(await vault.totalUnderlying()).to.be.equal(totalUnderlying);
+
+      amount = utils.parseUnits("1000", 6);
+      console.log(`Deposit ${utils.formatUnits(amount, 6)} USDT by bob`);
       await vault.connect(bob).deposit({
         amount,
         claims: [
@@ -156,21 +178,24 @@ describe("USTStrategy Mainnet fork", () => {
         lockedUntil: 0,
       });
 
+      console.log((await vault.totalUnderlying()).toString());
+      expect(await usdtToken.balanceOf(vault.address)).to.be.equal(amount);
       expect(await vault.totalUnderlying()).to.be.equal(
         totalUnderlying.add(amount)
       );
-      totalUnderlying = await vault.totalUnderlying();
+      totalUnderlying = totalUnderlying.add(amount);
 
       console.log(
-        `Invest: totalUnderlying - ${utils.formatEther(
-          await vault.totalUnderlying()
-        )} UST`
+        `Invest: totalUnderlying - ${utils.formatUnits(
+          totalUnderlying,
+          6
+        )} USDT`
       );
       exchangeRate = (await mockAUstUstFeed.latestRoundData()).answer;
       console.log("ExchangeRate: ", utils.formatEther(exchangeRate));
       await vault.updateInvested();
 
-      exchangeRate = await (await mockAUstUstFeed.latestRoundData()).answer;
+      exchangeRate = (await mockAUstUstFeed.latestRoundData()).answer;
       console.log("ExchangeRate: ", utils.formatEther(exchangeRate));
 
       depositOperations = await strategy.depositOperations(0);
@@ -191,15 +216,23 @@ describe("USTStrategy Mainnet fork", () => {
       );
 
       console.log(
-        `FinishDepositStable: totalUnderlying - ${utils.formatEther(
-          await vault.totalUnderlying()
-        )} UST (0 UST + ${utils.formatEther(aUstBalance)} aUST)`
+        `FinishDepositStable: totalUnderlying - ${utils.formatUnits(
+          await vault.totalUnderlying(),
+          6
+        )} USDT (${utils.formatUnits(
+          await usdtToken.balanceOf(strategy.address),
+          6
+        )} USDT + ${utils.formatEther(
+          await ustToken.balanceOf(strategy.address)
+        )} UST + ${utils.formatEther(
+          await aUstToken.balanceOf(strategy.address)
+        )} aUST)`
       );
     });
   });
 
   // Use Mock aUST / UST Chainlink Feed to check performance fee and redeem
-  describe("Use Mock aUST / UST Chainlink feed", () => {
+  describe("Use Mock aUST / UST Chainlink feed (Underlying: USDT)", () => {
     before(async () => {
       const MockChainlinkPriceFeedFactory = await ethers.getContractFactory(
         "MockChainlinkPriceFeed"
@@ -208,15 +241,17 @@ describe("USTStrategy Mainnet fork", () => {
 
       const VaultFactory = await ethers.getContractFactory("Vault");
       vault = await VaultFactory.deploy(
-        ustToken.address,
+        usdtToken.address,
         MIN_LOCK_PERIOD,
         INVEST_PCT,
         owner.address
       );
 
-      const USTStrategyFactory = await ethers.getContractFactory("USTStrategy");
+      const AnchorNonUSTStrategyFactory = await ethers.getContractFactory(
+        "AnchorNonUSTStrategy"
+      );
 
-      strategy = await USTStrategyFactory.deploy(
+      strategy = await AnchorNonUSTStrategyFactory.deploy(
         vault.address,
         TREASURY,
         config.ethAnchorRouter,
@@ -224,23 +259,28 @@ describe("USTStrategy Mainnet fork", () => {
         ustToken.address,
         aUstToken.address,
         FEE_PCT,
-        owner.address
+        owner.address,
+        config.curve,
+        config.usdtI,
+        config.ustI
       );
 
       await strategy.connect(owner).grantRole(MANAGER_ROLE, owner.address);
 
+      await strategy.initializeStrategy(config.ustFeed, config.usdtFeed);
+
       await vault.setStrategy(strategy.address);
 
-      await ustToken
+      await usdtToken
         .connect(alice)
         .approve(vault.address, constants.MaxUint256);
-      await ustToken.connect(bob).approve(vault.address, constants.MaxUint256);
+      await usdtToken.connect(bob).approve(vault.address, constants.MaxUint256);
     });
 
     it("Mainnet fork (2)", async () => {
-      let amount = utils.parseEther("9000");
+      let amount = utils.parseUnits("9000", 6);
 
-      console.log(`Deposit ${utils.formatEther(amount)} UST by alice`);
+      console.log(`Deposit ${utils.formatUnits(amount, 6)} USDT by alice`);
       await vault.connect(alice).deposit({
         amount,
         claims: [
@@ -252,22 +292,34 @@ describe("USTStrategy Mainnet fork", () => {
         ],
         lockedUntil: 0,
       });
-      expect(await ustToken.balanceOf(vault.address)).to.be.equal(amount);
+      expect(await usdtToken.balanceOf(vault.address)).to.be.equal(amount);
       let exchangeRate = utils.parseEther("1.17");
       await mockAUstUstFeed.setAnswer(exchangeRate);
 
       console.log("ExchangeRate: ", utils.formatEther(exchangeRate));
 
       console.log(
-        `Invest: totalUnderlying - ${utils.formatEther(
-          await vault.totalUnderlying()
-        )} UST`
+        `Invest: totalUnderlying - ${utils.formatUnits(
+          await vault.totalUnderlying(),
+          6
+        )} USDT`
       );
       await vault.connect(owner).updateInvested();
-      expect(await ustToken.balanceOf(vault.address)).to.be.equal("0");
-      expect(await ustToken.balanceOf(strategy.address)).to.be.equal("0");
-      expect(await strategy.pendingDeposits()).to.be.equal(amount);
-      expect(await vault.totalUnderlying()).to.be.equal(amount);
+      expect(await usdtToken.balanceOf(vault.address)).to.be.equal("0");
+      expect(await usdtToken.balanceOf(strategy.address)).to.be.equal("0");
+      expect(await strategy.pendingDeposits()).to.be.equal(
+        "8986216514701659263596" // raw data in fork block
+      );
+      console.log(
+        `Invest after: totalUnderlying - ${utils.formatUnits(
+          await vault.totalUnderlying(),
+          6
+        )} USDT`
+      );
+
+      expect(await vault.totalUnderlying()).to.be.equal(
+        "8993877694" // raw data in fork block
+      );
 
       let depositOperations = await strategy.depositOperations(0);
       let expectAUstReceive = utils.parseEther("7600");
@@ -286,13 +338,20 @@ describe("USTStrategy Mainnet fork", () => {
       let totalUnderlying = await vault.totalUnderlying();
 
       console.log(
-        `FinishDepositStable: totalUnderlying - ${utils.formatEther(
-          totalUnderlying
-        )} UST (0 UST + ${utils.formatEther(expectAUstReceive)} aUST)`
+        `Converted UST - ${utils.formatEther(
+          await strategy.convertedUst()
+        )} UST`
       );
 
-      amount = utils.parseEther("1000");
-      console.log(`Deposit ${utils.formatEther(amount)} UST by bob`);
+      console.log(
+        `FinishDepositStable: totalUnderlying - ${utils.formatUnits(
+          totalUnderlying,
+          6
+        )} USDT (0 USDT + 0 UST + ${utils.formatEther(expectAUstReceive)} aUST)`
+      );
+
+      amount = utils.parseUnits("1000", 6);
+      console.log(`Deposit ${utils.formatUnits(amount, 6)} USDT by bob`);
       await vault.connect(bob).deposit({
         amount,
         claims: [
@@ -335,9 +394,16 @@ describe("USTStrategy Mainnet fork", () => {
       );
 
       console.log(
-        `FinishDepositStable: totalUnderlying - ${utils.formatEther(
-          await vault.totalUnderlying()
-        )} UST (0 UST + ${utils.formatEther(aUstBalance)} aUST)`
+        `FinishDepositStable: totalUnderlying - ${utils.formatUnits(
+          await vault.totalUnderlying(),
+          6
+        )} USDT (0 USDT + ${utils.formatEther(aUstBalance)} aUST)`
+      );
+
+      console.log(
+        `Converted UST - ${utils.formatEther(
+          await strategy.convertedUst()
+        )} UST`
       );
 
       exchangeRate = utils.parseEther("1.3");
@@ -349,14 +415,16 @@ describe("USTStrategy Mainnet fork", () => {
       );
 
       console.log(
-        `TotalUnderlying - ${utils.formatEther(
-          await vault.totalUnderlying()
-        )} UST`
+        `TotalUnderlying - ${utils.formatUnits(
+          await vault.totalUnderlying(),
+          6
+        )} USDT`
       );
 
       totalUnderlying = await vault.totalUnderlying();
 
-      // profit: 985 UST
+      let convertedUst = await strategy.convertedUst();
+
       let redeemAmount = utils.parseEther("5000");
       await strategy.initRedeemStable(redeemAmount);
       expect(await strategy.pendingRedeems()).to.be.equal(redeemAmount);
@@ -371,25 +439,20 @@ describe("USTStrategy Mainnet fork", () => {
       );
 
       await strategy.finishRedeemStable(0);
-      let originalDeposit = utils
-        .parseEther("10000")
-        .mul(redeemAmount)
-        .div(aUstBalance);
+      let originalDeposit = convertedUst.mul(redeemAmount).div(aUstBalance);
       let profit = expectUstReceive.sub(originalDeposit);
       let fee = profit.mul(FEE_PCT).div(DENOMINATOR);
       expect(await ustToken.balanceOf(TREASURY)).to.be.equal(fee);
       console.log(
         `Finish redeem stable: profit - ${utils.formatEther(profit)} UST`
       );
-      expect(await vault.totalUnderlying()).to.be.equal(
-        "10867300000000000000001"
-      );
+      expect(await vault.totalUnderlying()).to.be.equal("10873546114");
       aUstBalance = aUstBalance.sub(redeemAmount);
       expect(await aUstToken.balanceOf(strategy.address)).to.be.equal(
         aUstBalance
       );
-      expect(await ustToken.balanceOf(vault.address)).to.be.equal(
-        expectUstReceive.sub(fee)
+      expect(await usdtToken.balanceOf(vault.address)).to.be.equal(
+        "6392897697"
       );
     });
   });
