+++ test/strategy/anchor/fork/AnchorUSTStrategy.fork.test.ts
@@ -4,27 +4,24 @@ import { expect } from "chai";
 import { BigNumber, utils, constants } from "ethers";
 import {
   Vault,
-  NonUSTStrategy,
+  AnchorUSTStrategy,
   MockERC20,
   MockChainlinkPriceFeed,
   MockERC20__factory,
   MockChainlinkPriceFeed__factory,
-} from "../../../typechain";
-import { generateNewAddress, ForkHelpers } from "../../shared/";
+} from "../../../../typechain";
+import { generateNewAddress, ForkHelpers } from "../../../shared";
 import config from "./config.json";
 
-describe("NonUSTStrategy Mainnet fork", () => {
+describe("AnchorUSTStrategy Mainnet fork", () => {
   let owner: SignerWithAddress;
   let alice: SignerWithAddress;
   let bob: SignerWithAddress;
 
   let vault: Vault;
-  let strategy: NonUSTStrategy;
+  let strategy: AnchorUSTStrategy;
   let ustToken: MockERC20;
   let aUstToken: MockERC20;
-  let usdtToken: MockERC20;
-  let usdcToken: MockERC20;
-  let daiToken: MockERC20;
   // MockChainlinkPriceFeed has same interface as Mainnet, so we can use it for test
   let mockAUstUstFeed: MockChainlinkPriceFeed;
   const MIN_LOCK_PERIOD = 0; // set zero for test
@@ -39,25 +36,22 @@ describe("NonUSTStrategy Mainnet fork", () => {
     await ForkHelpers.forkToMainnet(FORK_BLOCK);
     [owner, alice, bob] = await ethers.getSigners();
 
-    daiToken = new MockERC20__factory(owner).attach(config.dai);
-    usdcToken = new MockERC20__factory(owner).attach(config.usdc);
-    usdtToken = new MockERC20__factory(owner).attach(config.usdt);
     ustToken = new MockERC20__factory(owner).attach(config.ust);
     aUstToken = new MockERC20__factory(owner).attach(config.aUst);
 
     await ForkHelpers.mintToken(
-      usdtToken,
+      ustToken,
       alice,
-      utils.parseUnits("100000000000", 6)
+      utils.parseEther("100000000000")
     );
     await ForkHelpers.mintToken(
-      usdtToken,
+      ustToken,
       bob,
-      utils.parseUnits("100000000000", 6)
+      utils.parseEther("100000000000")
     );
   });
 
-  describe("Use Mainnet aUST / UST Chainlink feed (Underlying: USDT)", () => {
+  describe("Use Mainnet aUST / UST Chainlink feed", () => {
     before(async () => {
       mockAUstUstFeed = new MockChainlinkPriceFeed__factory(owner).attach(
         config.aUstUstFeed
@@ -65,17 +59,17 @@ describe("NonUSTStrategy Mainnet fork", () => {
 
       const VaultFactory = await ethers.getContractFactory("Vault");
       vault = await VaultFactory.deploy(
-        usdtToken.address,
+        ustToken.address,
         MIN_LOCK_PERIOD,
         INVEST_PCT,
         owner.address
       );
 
-      const NonUSTStrategyFactory = await ethers.getContractFactory(
-        "NonUSTStrategy"
+      const AnchorUSTStrategyFactory = await ethers.getContractFactory(
+        "AnchorUSTStrategy"
       );
 
-      strategy = await NonUSTStrategyFactory.deploy(
+      strategy = await AnchorUSTStrategyFactory.deploy(
         vault.address,
         TREASURY,
         config.ethAnchorRouter,
@@ -83,28 +77,23 @@ describe("NonUSTStrategy Mainnet fork", () => {
         ustToken.address,
         aUstToken.address,
         FEE_PCT,
-        owner.address,
-        config.curve,
-        config.usdtI,
-        config.ustI
+        owner.address
       );
 
       await strategy.connect(owner).grantRole(MANAGER_ROLE, owner.address);
 
-      await strategy.initializeStrategy(config.ustFeed, config.usdtFeed);
-
       await vault.setStrategy(strategy.address);
 
-      await usdtToken
+      await ustToken
         .connect(alice)
         .approve(vault.address, constants.MaxUint256);
-      await usdtToken.connect(bob).approve(vault.address, constants.MaxUint256);
+      await ustToken.connect(bob).approve(vault.address, constants.MaxUint256);
     });
 
     it("Mainnet fork (1)", async () => {
-      let amount = utils.parseUnits("10000", 6);
+      let amount = utils.parseEther("10000");
 
-      console.log(`Deposit ${utils.formatUnits(amount, 6)} USDT by alice`);
+      console.log(`Deposit ${utils.formatEther(amount)} UST by alice`);
       await vault.connect(alice).deposit({
         amount,
         claims: [
@@ -116,26 +105,20 @@ describe("NonUSTStrategy Mainnet fork", () => {
         ],
         lockedUntil: 0,
       });
-      expect(await usdtToken.balanceOf(vault.address)).to.be.equal(amount);
+      expect(await ustToken.balanceOf(vault.address)).to.be.equal(amount);
       let exchangeRate = (await mockAUstUstFeed.latestRoundData()).answer;
       console.log("ExchangeRate: ", utils.formatEther(exchangeRate));
 
-      await vault.connect(owner).updateInvested();
       console.log(
-        `Invest: totalUnderlying - ${utils.formatUnits(
-          await vault.totalUnderlying(),
-          6
-        )} USDT`
+        `Invest: totalUnderlying - ${utils.formatEther(
+          await vault.totalUnderlying()
+        )} UST`
       );
-
+      await vault.connect(owner).updateInvested();
       expect(await ustToken.balanceOf(vault.address)).to.be.equal("0");
       expect(await ustToken.balanceOf(strategy.address)).to.be.equal("0");
-      expect(await usdtToken.balanceOf(vault.address)).to.be.equal("0");
-      expect(await usdtToken.balanceOf(strategy.address)).to.be.equal("0");
-      expect(await strategy.pendingDeposits()).to.be.equal(
-        "9984797021624030505513" // raw data in fork block
-      );
-      let totalUnderlying = BigNumber.from(await vault.totalUnderlying());
+      expect(await strategy.pendingDeposits()).to.be.equal(amount);
+      expect(await vault.totalUnderlying()).to.be.equal(amount);
 
       let depositOperations = await strategy.depositOperations(0);
       let expectAUstReceive = utils.parseEther("8500");
@@ -151,21 +134,18 @@ describe("NonUSTStrategy Mainnet fork", () => {
         expectAUstReceive
       );
 
+      let totalUnderlying = await vault.totalUnderlying();
       exchangeRate = (await mockAUstUstFeed.latestRoundData()).answer;
       console.log("ExchangeRate: ", utils.formatEther(exchangeRate));
 
-      totalUnderlying = await vault.totalUnderlying();
       console.log(
-        `FinishDepositStable: totalUnderlying - ${utils.formatUnits(
-          totalUnderlying,
-          6
-        )} USDT (0 USDT + 0 UST + ${utils.formatEther(expectAUstReceive)} aUST)`
+        `FinishDepositStable: totalUnderlying - ${utils.formatEther(
+          totalUnderlying
+        )} UST (0 UST + ${utils.formatEther(expectAUstReceive)} aUST)`
       );
 
-      expect(await vault.totalUnderlying()).to.be.equal(totalUnderlying);
-
-      amount = utils.parseUnits("1000", 6);
-      console.log(`Deposit ${utils.formatUnits(amount, 6)} USDT by bob`);
+      amount = utils.parseEther("1000");
+      console.log(`Deposit ${utils.formatEther(amount)} UST by bob`);
       await vault.connect(bob).deposit({
         amount,
         claims: [
@@ -178,24 +158,21 @@ describe("NonUSTStrategy Mainnet fork", () => {
         lockedUntil: 0,
       });
 
-      console.log((await vault.totalUnderlying()).toString());
-      expect(await usdtToken.balanceOf(vault.address)).to.be.equal(amount);
       expect(await vault.totalUnderlying()).to.be.equal(
         totalUnderlying.add(amount)
       );
-      totalUnderlying = totalUnderlying.add(amount);
+      totalUnderlying = await vault.totalUnderlying();
 
       console.log(
-        `Invest: totalUnderlying - ${utils.formatUnits(
-          totalUnderlying,
-          6
-        )} USDT`
+        `Invest: totalUnderlying - ${utils.formatEther(
+          await vault.totalUnderlying()
+        )} UST`
       );
       exchangeRate = (await mockAUstUstFeed.latestRoundData()).answer;
       console.log("ExchangeRate: ", utils.formatEther(exchangeRate));
       await vault.updateInvested();
 
-      exchangeRate = (await mockAUstUstFeed.latestRoundData()).answer;
+      exchangeRate = await (await mockAUstUstFeed.latestRoundData()).answer;
       console.log("ExchangeRate: ", utils.formatEther(exchangeRate));
 
       depositOperations = await strategy.depositOperations(0);
@@ -216,23 +193,15 @@ describe("NonUSTStrategy Mainnet fork", () => {
       );
 
       console.log(
-        `FinishDepositStable: totalUnderlying - ${utils.formatUnits(
-          await vault.totalUnderlying(),
-          6
-        )} USDT (${utils.formatUnits(
-          await usdtToken.balanceOf(strategy.address),
-          6
-        )} USDT + ${utils.formatEther(
-          await ustToken.balanceOf(strategy.address)
-        )} UST + ${utils.formatEther(
-          await aUstToken.balanceOf(strategy.address)
-        )} aUST)`
+        `FinishDepositStable: totalUnderlying - ${utils.formatEther(
+          await vault.totalUnderlying()
+        )} UST (0 UST + ${utils.formatEther(aUstBalance)} aUST)`
       );
     });
   });
 
   // Use Mock aUST / UST Chainlink Feed to check performance fee and redeem
-  describe("Use Mock aUST / UST Chainlink feed (Underlying: USDT)", () => {
+  describe("Use Mock aUST / UST Chainlink feed", () => {
     before(async () => {
       const MockChainlinkPriceFeedFactory = await ethers.getContractFactory(
         "MockChainlinkPriceFeed"
@@ -241,17 +210,17 @@ describe("NonUSTStrategy Mainnet fork", () => {
 
       const VaultFactory = await ethers.getContractFactory("Vault");
       vault = await VaultFactory.deploy(
-        usdtToken.address,
+        ustToken.address,
         MIN_LOCK_PERIOD,
         INVEST_PCT,
         owner.address
       );
 
-      const NonUSTStrategyFactory = await ethers.getContractFactory(
-        "NonUSTStrategy"
+      const AnchorUSTStrategyFactory = await ethers.getContractFactory(
+        "AnchorUSTStrategy"
       );
 
-      strategy = await NonUSTStrategyFactory.deploy(
+      strategy = await AnchorUSTStrategyFactory.deploy(
         vault.address,
         TREASURY,
         config.ethAnchorRouter,
@@ -259,28 +228,23 @@ describe("NonUSTStrategy Mainnet fork", () => {
         ustToken.address,
         aUstToken.address,
         FEE_PCT,
-        owner.address,
-        config.curve,
-        config.usdtI,
-        config.ustI
+        owner.address
       );
 
       await strategy.connect(owner).grantRole(MANAGER_ROLE, owner.address);
 
-      await strategy.initializeStrategy(config.ustFeed, config.usdtFeed);
-
       await vault.setStrategy(strategy.address);
 
-      await usdtToken
+      await ustToken
         .connect(alice)
         .approve(vault.address, constants.MaxUint256);
-      await usdtToken.connect(bob).approve(vault.address, constants.MaxUint256);
+      await ustToken.connect(bob).approve(vault.address, constants.MaxUint256);
     });
 
     it("Mainnet fork (2)", async () => {
-      let amount = utils.parseUnits("9000", 6);
+      let amount = utils.parseEther("9000");
 
-      console.log(`Deposit ${utils.formatUnits(amount, 6)} USDT by alice`);
+      console.log(`Deposit ${utils.formatEther(amount)} UST by alice`);
       await vault.connect(alice).deposit({
         amount,
         claims: [
@@ -292,34 +256,22 @@ describe("NonUSTStrategy Mainnet fork", () => {
         ],
         lockedUntil: 0,
       });
-      expect(await usdtToken.balanceOf(vault.address)).to.be.equal(amount);
+      expect(await ustToken.balanceOf(vault.address)).to.be.equal(amount);
       let exchangeRate = utils.parseEther("1.17");
       await mockAUstUstFeed.setAnswer(exchangeRate);
 
       console.log("ExchangeRate: ", utils.formatEther(exchangeRate));
 
       console.log(
-        `Invest: totalUnderlying - ${utils.formatUnits(
-          await vault.totalUnderlying(),
-          6
-        )} USDT`
+        `Invest: totalUnderlying - ${utils.formatEther(
+          await vault.totalUnderlying()
+        )} UST`
       );
       await vault.connect(owner).updateInvested();
-      expect(await usdtToken.balanceOf(vault.address)).to.be.equal("0");
-      expect(await usdtToken.balanceOf(strategy.address)).to.be.equal("0");
-      expect(await strategy.pendingDeposits()).to.be.equal(
-        "8986216514701659263596" // raw data in fork block
-      );
-      console.log(
-        `Invest after: totalUnderlying - ${utils.formatUnits(
-          await vault.totalUnderlying(),
-          6
-        )} USDT`
-      );
-
-      expect(await vault.totalUnderlying()).to.be.equal(
-        "8993877694" // raw data in fork block
-      );
+      expect(await ustToken.balanceOf(vault.address)).to.be.equal("0");
+      expect(await ustToken.balanceOf(strategy.address)).to.be.equal("0");
+      expect(await strategy.pendingDeposits()).to.be.equal(amount);
+      expect(await vault.totalUnderlying()).to.be.equal(amount);
 
       let depositOperations = await strategy.depositOperations(0);
       let expectAUstReceive = utils.parseEther("7600");
@@ -338,20 +290,13 @@ describe("NonUSTStrategy Mainnet fork", () => {
       let totalUnderlying = await vault.totalUnderlying();
 
       console.log(
-        `Converted UST - ${utils.formatEther(
-          await strategy.convertedUst()
-        )} UST`
+        `FinishDepositStable: totalUnderlying - ${utils.formatEther(
+          totalUnderlying
+        )} UST (0 UST + ${utils.formatEther(expectAUstReceive)} aUST)`
       );
 
-      console.log(
-        `FinishDepositStable: totalUnderlying - ${utils.formatUnits(
-          totalUnderlying,
-          6
-        )} USDT (0 USDT + 0 UST + ${utils.formatEther(expectAUstReceive)} aUST)`
-      );
-
-      amount = utils.parseUnits("1000", 6);
-      console.log(`Deposit ${utils.formatUnits(amount, 6)} USDT by bob`);
+      amount = utils.parseEther("1000");
+      console.log(`Deposit ${utils.formatEther(amount)} UST by bob`);
       await vault.connect(bob).deposit({
         amount,
         claims: [
@@ -394,16 +339,9 @@ describe("NonUSTStrategy Mainnet fork", () => {
       );
 
       console.log(
-        `FinishDepositStable: totalUnderlying - ${utils.formatUnits(
-          await vault.totalUnderlying(),
-          6
-        )} USDT (0 USDT + ${utils.formatEther(aUstBalance)} aUST)`
-      );
-
-      console.log(
-        `Converted UST - ${utils.formatEther(
-          await strategy.convertedUst()
-        )} UST`
+        `FinishDepositStable: totalUnderlying - ${utils.formatEther(
+          await vault.totalUnderlying()
+        )} UST (0 UST + ${utils.formatEther(aUstBalance)} aUST)`
       );
 
       exchangeRate = utils.parseEther("1.3");
@@ -415,16 +353,14 @@ describe("NonUSTStrategy Mainnet fork", () => {
       );
 
       console.log(
-        `TotalUnderlying - ${utils.formatUnits(
-          await vault.totalUnderlying(),
-          6
-        )} USDT`
+        `TotalUnderlying - ${utils.formatEther(
+          await vault.totalUnderlying()
+        )} UST`
       );
 
       totalUnderlying = await vault.totalUnderlying();
 
-      let convertedUst = await strategy.convertedUst();
-
+      // profit: 985 UST
       let redeemAmount = utils.parseEther("5000");
       await strategy.initRedeemStable(redeemAmount);
       expect(await strategy.pendingRedeems()).to.be.equal(redeemAmount);
@@ -439,20 +375,25 @@ describe("NonUSTStrategy Mainnet fork", () => {
       );
 
       await strategy.finishRedeemStable(0);
-      let originalDeposit = convertedUst.mul(redeemAmount).div(aUstBalance);
+      let originalDeposit = utils
+        .parseEther("10000")
+        .mul(redeemAmount)
+        .div(aUstBalance);
       let profit = expectUstReceive.sub(originalDeposit);
       let fee = profit.mul(FEE_PCT).div(DENOMINATOR);
       expect(await ustToken.balanceOf(TREASURY)).to.be.equal(fee);
       console.log(
         `Finish redeem stable: profit - ${utils.formatEther(profit)} UST`
       );
-      expect(await vault.totalUnderlying()).to.be.equal("10873546114");
+      expect(await vault.totalUnderlying()).to.be.equal(
+        "10867300000000000000001"
+      );
       aUstBalance = aUstBalance.sub(redeemAmount);
       expect(await aUstToken.balanceOf(strategy.address)).to.be.equal(
         aUstBalance
       );
-      expect(await usdtToken.balanceOf(vault.address)).to.be.equal(
-        "6392897697"
+      expect(await ustToken.balanceOf(vault.address)).to.be.equal(
+        expectUstReceive.sub(fee)
       );
     });
   });
